requirements.txt
Flask==2.1.1
slack-sdk==3.13.0
discord.py==1.7.3
aws-wsgi==0.2.7

